moved types for overridable component to @material-ui/types
